Sales history service implemented

+++ src/main/java/com/prodnretail/inventory_gen/service/CategoryService.java
@@ -1,65 +1,61 @@
 package com.prodnretail.inventory_gen.service;
 
 import com.prodnretail.inventory_gen.dto.CategoryDTO;
+import com.prodnretail.inventory_gen.exception.ResourceNotFoundException;
 import com.prodnretail.inventory_gen.model.Category;
 import com.prodnretail.inventory_gen.repository.CategoryRepository;
+
+import lombok.AllArgsConstructor;
+
 import org.springframework.stereotype.Service;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 import org.modelmapper.ModelMapper;
 @Service
+@AllArgsConstructor
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
     private final ModelMapper modelMapper;
-    public CategoryService(CategoryRepository categoryRepository, ModelMapper modelMapper) {
-        this.categoryRepository = categoryRepository;
-        this.modelMapper=modelMapper;
+
+    
+    public List<CategoryDTO> getAll() {
+        return categoryRepository.findAll()
+        .stream()
+        .map(category -> modelMapper.map(category,CategoryDTO.class)).toList();
     }
 
-    public Category saveCategory(Category category){
-        return categoryRepository.save(category);
-    }
-    public List<Category> getAllCategories() {
-        return categoryRepository.findAll();
-    }
-
-    public Optional<Category> getCategoryById(UUID id) {
-        return categoryRepository.findById(id);
-    }
-
-    public Category createCategoryFromDto(CategoryDTO dto) {
+    public CategoryDTO create(CategoryDTO dto) {
         Category category = modelMapper.map(dto, Category.class);
-
-        return categoryRepository.save(category);
+        Category saved = categoryRepository.save(category);
+        return modelMapper.map(saved,CategoryDTO.class);
 
     }
+    public CategoryDTO getById(UUID id){
+        Category category = categoryRepository.findById(id)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("Category not found with id: " + id)
+                );
+        return modelMapper.map(category, CategoryDTO.class);
+    }
 
-    public Category updateCategoryFromDto(UUID id, CategoryDTO dto) {
-        return categoryRepository.findById(id).map(existing -> {
+    public CategoryDTO update(UUID id, CategoryDTO dto) {
+        Category existing = categoryRepository.findById(id)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("Category not found with id: " + id)
+                );
             modelMapper.map(dto,existing);
+            Category updated =categoryRepository.save(existing);
 
-            return categoryRepository.save(existing);
-        }).orElseThrow(() -> new RuntimeException("Category not found with id: " + id));
+            return modelMapper.map(updated, CategoryDTO.class);
     }
 
-    public void deleteCategory(UUID id) {
+    public void delete(UUID id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Category not found with id: " + id);
+        }
         categoryRepository.deleteById(id);
     }
 
-    public Category toEntity(CategoryDTO dto){
-        Category category= new Category();
-        category.setId(dto.getId());
-        category.setCategoryName(dto.getCategoryName());
-        return category;
-    }
-
-    public CategoryDTO toDto(Category category){
-        return CategoryDTO.builder()
-        .id(category.getId())
-        .categoryName(category.getCategoryName())
-        .build();
-    }
 }
